MR B4: Decompose md2pdf.py into package (#122)
* refactor: decompose md2pdf.py into package (MR B4)

Decompose the 767-line md2pdf monolith into a clean package:

- openpaw/builtins/tools/md2pdf/__init__.py     Md2pdfToolBuiltin facade (~217 lines)
- openpaw/builtins/tools/md2pdf/models.py       Md2pdfInput, _MermaidBlock, RepairState (~56 lines)
- openpaw/builtins/tools/md2pdf/mermaid.py      Mermaid rendering, SVG scaling, self-healing (~362 lines)
- openpaw/builtins/tools/md2pdf/converter.py   Markdown-to-HTML-to-PDF pipeline (~163 lines)

Tests split into focused modules:
- tests/builtins/md2pdf/test_mermaid.py   Extraction, scaling, rendering, self-heal
- tests/builtins/md2pdf/test_converter.py Markdown-to-HTML, result messages
- tests/builtins/md2pdf/test_builtin.py   Builtin class, metadata, tool function, registry

All 66 md2pdf tests pass. Full suite: 2944 passed. Ruff clean.

* refactor: address AI review feedback for MR B4

- mermaid.py: catch httpx.HTTPError base class for explicit network error handling
- converter.py: narrow exception handling to (OSError, UnicodeDecodeError) for file reads

* docs: update session handoff for MR B4

---------
