+++ tests/builtins/md2pdf/test_mermaid.py
@@ -1,8 +1,6 @@
-"""Tests for the md2pdf builtin tool.
+"""Tests for md2pdf mermaid rendering and self-healing.
 
-Covers pure helper functions, mocked network/AI calls, and the
-Md2pdfToolBuiltin class itself.  WeasyPrint is always mocked because it
-is a heavy optional dependency that is not required to run the test suite.
+Covers pure helper functions, mocked network/AI calls, and SVG manipulation.
 """
 
 from pathlib import Path
@@ -11,12 +9,8 @@ from unittest.mock import MagicMock, patch
 import httpx
 import pytest
 
-from openpaw.builtins.base import BuiltinType
 from openpaw.builtins.tools.md2pdf import (
-    Md2pdfToolBuiltin,
-    _build_result_message,
     _extract_mermaid_blocks,
-    _markdown_to_html,
     _MermaidBlock,
     _render_all_blocks,
     _render_mermaid_to_svg,
@@ -24,40 +18,6 @@ from openpaw.builtins.tools.md2pdf import (
     _scale_svg_to_width,
     _try_self_heal,
 )
-from openpaw.builtins.tools.md2pdf_themes import DEFAULT_THEME, THEMES
-
-# ---------------------------------------------------------------------------
-# Fixtures
-# ---------------------------------------------------------------------------
-
-
-@pytest.fixture
-def workspace(tmp_path: Path) -> Path:
-    """Return a temporary workspace directory."""
-    ws = tmp_path / "my_agent"
-    ws.mkdir()
-    return ws
-
-
-@pytest.fixture
-def source_md(workspace: Path) -> Path:
-    """Write a minimal markdown file and return its path."""
-    md_file = workspace / "report.md"
-    md_file.write_text("# Hello\n\nSome content here.\n", encoding="utf-8")
-    return md_file
-
-
-@pytest.fixture
-def tool(workspace: Path) -> Md2pdfToolBuiltin:
-    """Return an Md2pdfToolBuiltin wired to the temp workspace."""
-    return Md2pdfToolBuiltin(config={"workspace_path": str(workspace)})
-
-
-@pytest.fixture
-def tool_fn(tool: Md2pdfToolBuiltin):
-    """Return the inner callable from get_langchain_tool()."""
-    return tool.get_langchain_tool().func
-
 
 # ---------------------------------------------------------------------------
 # Helpers
@@ -230,89 +190,6 @@ class TestReplaceMermaidBlocks:
 
 
 # ---------------------------------------------------------------------------
-# TestMarkdownToHtml
-# ---------------------------------------------------------------------------
-
-
-class TestMarkdownToHtml:
-    def test_basic_heading_converted(self) -> None:
-        html = _markdown_to_html("# Hello World")
-        assert "<h1" in html
-        assert "Hello World" in html
-
-    def test_table_converted(self) -> None:
-        md = "| A | B |\n|---|---|\n| 1 | 2 |"
-        html = _markdown_to_html(md)
-        assert "<table" in html
-        assert "<th" in html
-
-    def test_fenced_code_block_converted(self) -> None:
-        md = "```python\nprint('hello')\n```"
-        html = _markdown_to_html(md)
-        assert "<code" in html
-
-    def test_empty_string_returns_empty_string(self) -> None:
-        html = _markdown_to_html("")
-        assert html == ""
-
-    def test_inline_svg_passes_through(self) -> None:
-        svg = _simple_svg()
-        html = _markdown_to_html(svg)
-        assert "<svg" in html
-
-
-# ---------------------------------------------------------------------------
-# TestBuildResultMessage
-# ---------------------------------------------------------------------------
-
-
-class TestBuildResultMessage:
-    def test_no_diagrams(self) -> None:
-        path = Path("/workspace/report.pdf")
-        msg = _build_result_message(path, [])
-        assert "PDF created" in msg
-        assert str(path) in msg
-        # No diagram stats when there are no blocks
-        assert "Mermaid" not in msg
-
-    def test_all_rendered_ok(self) -> None:
-        block = _make_block("flowchart LR\n  A --> B")
-        block.svg = _simple_svg()
-        msg = _build_result_message(Path("/out/report.pdf"), [block])
-        assert "1 total" in msg
-        assert "rendered successfully" in msg
-        assert "repaired" not in msg
-        assert "failed" not in msg
-
-    def test_mixed_results(self) -> None:
-        ok_block = _make_block("flowchart LR\n  A --> B")
-        ok_block.svg = _simple_svg()
-
-        repaired_block = _make_block("sequenceDiagram\n  Alice->>Bob: Hi")
-        repaired_block.svg = _simple_svg()
-        repaired_block.ai_repaired = True
-
-        failed_block = _make_block("stateDiagram-v2\n  [*] --> S1")
-        failed_block.error = "HTTP 400"
-
-        blocks = [ok_block, repaired_block, failed_block]
-        msg = _build_result_message(Path("/out/report.pdf"), blocks)
-
-        assert "3 total" in msg
-        assert "1 rendered successfully" in msg
-        assert "1 repaired by AI" in msg
-        assert "1 failed" in msg
-
-    def test_only_failed(self) -> None:
-        block = _make_block("bad diagram")
-        block.error = "Parse error"
-        msg = _build_result_message(Path("/out/report.pdf"), [block])
-        assert "failed" in msg
-        assert "rendered successfully" not in msg
-        assert "repaired" not in msg
-
-
-# ---------------------------------------------------------------------------
 # TestRenderMermaidToSvg
 # ---------------------------------------------------------------------------
 
@@ -323,7 +200,7 @@ class TestRenderMermaidToSvg:
         mock_response.text = _simple_svg(width=100, height=50)
         mock_response.raise_for_status = MagicMock()
 
-        with patch("openpaw.builtins.tools.md2pdf.httpx.get", return_value=mock_response):
+        with patch("openpaw.builtins.tools.md2pdf.mermaid.httpx.get", return_value=mock_response):
             svg, error = _render_mermaid_to_svg("flowchart LR\n  A --> B")
 
         assert svg is not None
@@ -341,7 +218,7 @@ class TestRenderMermaidToSvg:
             response=mock_response,
         )
 
-        with patch("openpaw.builtins.tools.md2pdf.httpx.get", side_effect=http_error):
+        with patch("openpaw.builtins.tools.md2pdf.mermaid.httpx.get", side_effect=http_error):
             svg, error = _render_mermaid_to_svg("invalid diagram")
 
         assert svg is None
@@ -350,7 +227,7 @@ class TestRenderMermaidToSvg:
 
     def test_network_error_returns_error_message(self) -> None:
         with patch(
-            "openpaw.builtins.tools.md2pdf.httpx.get",
+            "openpaw.builtins.tools.md2pdf.mermaid.httpx.get",
             side_effect=httpx.ConnectError("Connection refused"),
         ):
             svg, error = _render_mermaid_to_svg("flowchart LR\n  A --> B")
@@ -365,7 +242,7 @@ class TestRenderMermaidToSvg:
         mock_response.text = wide_svg
         mock_response.raise_for_status = MagicMock()
 
-        with patch("openpaw.builtins.tools.md2pdf.httpx.get", return_value=mock_response):
+        with patch("openpaw.builtins.tools.md2pdf.mermaid.httpx.get", return_value=mock_response):
             svg, error = _render_mermaid_to_svg("flowchart LR\n  A --> B", max_width_px=500)
 
         assert error is None
@@ -409,7 +286,7 @@ class TestTrySelfHeal:
         }
 
         with patch("langchain.chat_models.init_chat_model", return_value=mock_llm), \
-             patch("openpaw.builtins.tools.md2pdf._build_repair_graph", return_value=mock_graph):
+             patch("openpaw.builtins.tools.md2pdf.mermaid._build_repair_graph", return_value=mock_graph):
             _try_self_heal(block, 624, "openai:gpt-4o-mini", 3, "full markdown context")
 
         assert block.svg is not None
@@ -442,7 +319,7 @@ class TestTrySelfHeal:
         mock_graph.invoke.side_effect = RuntimeError("Graph exploded")
 
         with patch("langchain.chat_models.init_chat_model", return_value=mock_llm), \
-             patch("openpaw.builtins.tools.md2pdf._build_repair_graph", return_value=mock_graph):
+             patch("openpaw.builtins.tools.md2pdf.mermaid._build_repair_graph", return_value=mock_graph):
             _try_self_heal(block, 624, "openai:gpt-4o-mini", 3, "context")
 
         assert block.svg is None
@@ -464,7 +341,7 @@ class TestTrySelfHeal:
         }
 
         with patch("langchain.chat_models.init_chat_model", return_value=mock_llm), \
-             patch("openpaw.builtins.tools.md2pdf._build_repair_graph", return_value=mock_graph):
+             patch("openpaw.builtins.tools.md2pdf.mermaid._build_repair_graph", return_value=mock_graph):
             _try_self_heal(block, 624, "openai:gpt-4o-mini", 3, "context")
 
         assert block.svg is None
@@ -486,7 +363,7 @@ class TestRenderAllBlocks:
         svg = _simple_svg()
 
         with patch(
-            "openpaw.builtins.tools.md2pdf._render_mermaid_to_svg",
+            "openpaw.builtins.tools.md2pdf.mermaid._render_mermaid_to_svg",
             return_value=(svg, None),
         ):
             _render_all_blocks(
@@ -501,7 +378,7 @@ class TestRenderAllBlocks:
         block = self._block()
 
         with patch(
-            "openpaw.builtins.tools.md2pdf._render_mermaid_to_svg",
+            "openpaw.builtins.tools.md2pdf.mermaid._render_mermaid_to_svg",
             return_value=(None, "HTTP 400"),
         ):
             _render_all_blocks(
@@ -516,9 +393,9 @@ class TestRenderAllBlocks:
         block = self._block()
 
         with patch(
-            "openpaw.builtins.tools.md2pdf._render_mermaid_to_svg",
+            "openpaw.builtins.tools.md2pdf.mermaid._render_mermaid_to_svg",
             return_value=(None, "Parse error"),
-        ), patch("openpaw.builtins.tools.md2pdf._try_self_heal") as mock_heal:
+        ), patch("openpaw.builtins.tools.md2pdf.mermaid._try_self_heal") as mock_heal:
             _render_all_blocks(
                 [block], 624, self_heal=True,
                 model_spec="gpt-4o-mini", max_iterations=3, markdown_context="ctx",
@@ -530,9 +407,9 @@ class TestRenderAllBlocks:
         block = self._block()
 
         with patch(
-            "openpaw.builtins.tools.md2pdf._render_mermaid_to_svg",
+            "openpaw.builtins.tools.md2pdf.mermaid._render_mermaid_to_svg",
             return_value=(None, "Parse error"),
-        ), patch("openpaw.builtins.tools.md2pdf._try_self_heal") as mock_heal:
+        ), patch("openpaw.builtins.tools.md2pdf.mermaid._try_self_heal") as mock_heal:
             _render_all_blocks(
                 [block], 624, self_heal=False,
                 model_spec="gpt-4o-mini", max_iterations=3, markdown_context="ctx",
@@ -544,9 +421,9 @@ class TestRenderAllBlocks:
         block = self._block()
 
         with patch(
-            "openpaw.builtins.tools.md2pdf._render_mermaid_to_svg",
+            "openpaw.builtins.tools.md2pdf.mermaid._render_mermaid_to_svg",
             return_value=(_simple_svg(), None),
-        ), patch("openpaw.builtins.tools.md2pdf._try_self_heal") as mock_heal:
+        ), patch("openpaw.builtins.tools.md2pdf.mermaid._try_self_heal") as mock_heal:
             _render_all_blocks(
                 [block], 624, self_heal=True,
                 model_spec="gpt-4o-mini", max_iterations=3, markdown_context="ctx",
@@ -565,7 +442,7 @@ class TestRenderAllBlocks:
             call_count += 1
             return result
 
-        with patch("openpaw.builtins.tools.md2pdf._render_mermaid_to_svg", side_effect=fake_render):
+        with patch("openpaw.builtins.tools.md2pdf.mermaid._render_mermaid_to_svg", side_effect=fake_render):
             _render_all_blocks(
                 blocks, 624, self_heal=False,
                 model_spec="gpt-4o-mini", max_iterations=3, markdown_context="",
@@ -577,232 +454,34 @@ class TestRenderAllBlocks:
 
 
 # ---------------------------------------------------------------------------
-# TestMd2pdfToolBuiltin
+# Fixtures
 # ---------------------------------------------------------------------------
 
 
-class TestMd2pdfToolBuiltin:
-    def test_metadata_name(self) -> None:
-        assert Md2pdfToolBuiltin.metadata.name == "md2pdf"
-
-    def test_metadata_group(self) -> None:
-        assert Md2pdfToolBuiltin.metadata.group == "document"
-
-    def test_metadata_builtin_type(self) -> None:
-        assert Md2pdfToolBuiltin.metadata.builtin_type == BuiltinType.TOOL
-
-    def test_metadata_display_name(self) -> None:
-        assert Md2pdfToolBuiltin.metadata.display_name == "Markdown to PDF"
-
-    def test_metadata_prerequisites_include_weasyprint(self) -> None:
-        assert "weasyprint" in Md2pdfToolBuiltin.metadata.prerequisites.packages
-
-    def test_default_config_values(self) -> None:
-        tool = Md2pdfToolBuiltin()
-        assert tool.theme == DEFAULT_THEME
-        assert tool.max_diagram_width == 6.5
-        assert tool.self_heal is True
-        assert tool.self_heal_model == "gpt-4o-mini"
-        assert tool.max_heal_iterations == 3
-        assert tool.workspace_path is None
-
-    def test_custom_config_applied(self, workspace: Path) -> None:
-        tool = Md2pdfToolBuiltin(
-            config={
-                "workspace_path": str(workspace),
-                "theme": "professional",
-                "max_diagram_width": 5.0,
-                "self_heal": False,
-                "self_heal_model": "anthropic:claude-haiku-4-5",
-                "max_heal_iterations": 1,
-            }
-        )
-        assert tool.theme == "professional"
-        assert tool.max_diagram_width == 5.0
-        assert tool.self_heal is False
-        assert tool.self_heal_model == "anthropic:claude-haiku-4-5"
-        assert tool.max_heal_iterations == 1
-        assert tool.workspace_path == workspace.resolve()
-
-    def test_get_langchain_tool_returns_structured_tool(self, tool: Md2pdfToolBuiltin) -> None:
-        from langchain_core.tools import StructuredTool
-
-        lt = tool.get_langchain_tool()
-        assert isinstance(lt, StructuredTool)
-
-    def test_langchain_tool_name(self, tool: Md2pdfToolBuiltin) -> None:
-        lt = tool.get_langchain_tool()
-        assert lt.name == "markdown_to_pdf"
-
-    def test_langchain_tool_has_description(self, tool: Md2pdfToolBuiltin) -> None:
-        lt = tool.get_langchain_tool()
-        assert lt.description
-        assert "markdown" in lt.description.lower()
-
-    def test_langchain_tool_schema_has_source_path(self, tool: Md2pdfToolBuiltin) -> None:
-        lt = tool.get_langchain_tool()
-        schema = lt.args_schema.model_json_schema()
-        assert "source_path" in schema["properties"]
-
-    def test_langchain_tool_schema_has_output_path(self, tool: Md2pdfToolBuiltin) -> None:
-        lt = tool.get_langchain_tool()
-        schema = lt.args_schema.model_json_schema()
-        assert "output_path" in schema["properties"]
-
-    def test_langchain_tool_schema_has_theme(self, tool: Md2pdfToolBuiltin) -> None:
-        lt = tool.get_langchain_tool()
-        schema = lt.args_schema.model_json_schema()
-        assert "theme" in schema["properties"]
+@pytest.fixture
+def workspace(tmp_path: Path) -> Path:
+    """Return a temporary workspace directory."""
+    ws = tmp_path / "my_agent"
+    ws.mkdir()
+    return ws
 
 
-# ---------------------------------------------------------------------------
-# TestToolFunction
-# ---------------------------------------------------------------------------
+@pytest.fixture
+def source_md(workspace: Path) -> Path:
+    """Write a minimal markdown file and return its path."""
+    md_file = workspace / "report.md"
+    md_file.write_text("# Hello\n\nSome content here.\n", encoding="utf-8")
+    return md_file
 
 
-class TestToolFunction:
-    """Test the inner markdown_to_pdf callable via get_langchain_tool().func."""
-
-    def _mock_convert_success(self, output_path: Path, blocks=None):
-        """Return a mock for _convert that writes a dummy PDF and succeeds."""
-        if blocks is None:
-            blocks = []
-
-        def fake_convert(source_path, output_path, css, max_width_px, self_heal, model_spec, max_iterations):
-            output_path.parent.mkdir(parents=True, exist_ok=True)
-            output_path.write_bytes(b"%PDF-1.4 fake")
-            msg = _build_result_message(output_path, blocks)
-            return True, msg, blocks
-
-        return fake_convert
-
-    def test_no_workspace_path_returns_error(self) -> None:
-        tool = Md2pdfToolBuiltin(config={})
-        fn = tool.get_langchain_tool().func
-        result = fn(source_path="report.md")
-        assert "[Error:" in result
-        assert "workspace_path" in result
-
-    def test_source_not_found_returns_error(self, tool: Md2pdfToolBuiltin, workspace: Path) -> None:
-        fn = tool.get_langchain_tool().func
-        result = fn(source_path="nonexistent.md")
-        assert "[Error:" in result
-        assert "not found" in result
-
-    def test_invalid_source_path_returns_error(self, tool: Md2pdfToolBuiltin) -> None:
-        fn = tool.get_langchain_tool().func
-        # Path traversal is rejected by sandbox
-        result = fn(source_path="../escape.md")
-        assert "[Error:" in result
-
-    def test_success_returns_workspace_relative_path(
-        self, tool: Md2pdfToolBuiltin, workspace: Path, source_md: Path
-    ) -> None:
-        fake_convert = self._mock_convert_success(workspace / "report.pdf")
-
-        with patch("openpaw.builtins.tools.md2pdf._convert", side_effect=fake_convert):
-            result = fn = tool.get_langchain_tool().func
-            result = fn(source_path="report.md")
-
-        assert "[Error:" not in result
-        assert "report.pdf" in result
-        # Must be workspace-relative, not absolute
-        assert str(workspace) not in result
-
-    def test_default_output_path_uses_pdf_extension(
-        self, tool: Md2pdfToolBuiltin, workspace: Path, source_md: Path
-    ) -> None:
-        captured = {}
-
-        def capturing_convert(source_path, output_path, **kwargs):
-            captured["output_path"] = output_path
-            output_path.parent.mkdir(parents=True, exist_ok=True)
-            output_path.write_bytes(b"%PDF fake")
-            return True, _build_result_message(output_path, []), []
-
-        with patch("openpaw.builtins.tools.md2pdf._convert", side_effect=capturing_convert):
-            fn = tool.get_langchain_tool().func
-            fn(source_path="report.md")
-
-        assert captured["output_path"].suffix == ".pdf"
-        assert captured["output_path"].stem == "report"
-
-    def test_invalid_theme_falls_back_to_default(
-        self, tool: Md2pdfToolBuiltin, workspace: Path, source_md: Path
-    ) -> None:
-        captured = {}
-
-        def capturing_convert(source_path, output_path, css, **kwargs):
-            captured["css"] = css
-            output_path.parent.mkdir(parents=True, exist_ok=True)
-            output_path.write_bytes(b"%PDF fake")
-            return True, _build_result_message(output_path, []), []
-
-        with patch("openpaw.builtins.tools.md2pdf._convert", side_effect=capturing_convert):
-            fn = tool.get_langchain_tool().func
-            fn(source_path="report.md", theme="nonexistent_theme")
-
-        assert captured["css"] == THEMES[DEFAULT_THEME]
-
-    def test_valid_theme_passed_to_convert(
-        self, tool: Md2pdfToolBuiltin, workspace: Path, source_md: Path
-    ) -> None:
-        captured = {}
-
-        def capturing_convert(source_path, output_path, css, **kwargs):
-            captured["css"] = css
-            output_path.parent.mkdir(parents=True, exist_ok=True)
-            output_path.write_bytes(b"%PDF fake")
-            return True, _build_result_message(output_path, []), []
-
-        with patch("openpaw.builtins.tools.md2pdf._convert", side_effect=capturing_convert):
-            fn = tool.get_langchain_tool().func
-            fn(source_path="report.md", theme="technical")
-
-        assert captured["css"] == THEMES["technical"]
-
-    def test_convert_failure_returns_error_string(
-        self, tool: Md2pdfToolBuiltin, workspace: Path, source_md: Path
-    ) -> None:
-        with patch(
-            "openpaw.builtins.tools.md2pdf._convert",
-            return_value=(False, "PDF generation failed: disk full", []),
-        ):
-            fn = tool.get_langchain_tool().func
-            result = fn(source_path="report.md")
-
-        assert "[Error:" in result
-        assert "PDF generation failed" in result
-
-    def test_invalid_output_path_returns_error(
-        self, tool: Md2pdfToolBuiltin, workspace: Path, source_md: Path
-    ) -> None:
-        fn = tool.get_langchain_tool().func
-        result = fn(source_path="report.md", output_path="../outside.pdf")
-        assert "[Error:" in result
+@pytest.fixture
+def tool(workspace: Path):
+    """Return an Md2pdfToolBuiltin wired to the temp workspace."""
+    from openpaw.builtins.tools.md2pdf import Md2pdfToolBuiltin
+    return Md2pdfToolBuiltin(config={"workspace_path": str(workspace)})
 
 
-# ---------------------------------------------------------------------------
-# TestRegistry
-# ---------------------------------------------------------------------------
-
-
-class TestRegistry:
-    def test_md2pdf_builtin_is_registered(self) -> None:
-        from openpaw.builtins.registry import BuiltinRegistry
-
-        BuiltinRegistry.reset()
-        registry = BuiltinRegistry.get_instance()
-
-        assert "md2pdf" in registry._tools
-        assert registry._tools["md2pdf"] is Md2pdfToolBuiltin
-
-    def test_md2pdf_in_document_group(self) -> None:
-        from openpaw.builtins.registry import BuiltinRegistry
-
-        BuiltinRegistry.reset()
-        registry = BuiltinRegistry.get_instance()
-
-        groups = registry._groups
-        assert "document" in groups
-        assert "md2pdf" in groups["document"]
+@pytest.fixture
+def tool_fn(tool):
+    """Return the inner callable from get_langchain_tool()."""
+    return tool.get_langchain_tool().func
